+++ plugins/modules/ibmi_user_compliance_check.py
@@ -134,15 +134,14 @@ author:
 EXAMPLES = r'''
 - name: Do user profile compliance check
   ibm.power_ibmi.ibmi_user_compliance_check:
-      users:
-        - 'ibmiuser1'
-        - 'ibmiuser2'
-        - 'ibmiuser3'
-      fields:
-        - {'name':'status', 'expect':['*enabled']}
-        - {'name':'NO_PASSWORD_INDICATOR', 'expect':['no']}
-        - {'name':'SPECIAL_AUTHORITIES', 'expect': ['*JOBCTL','*SAVSYS']}
-
+    users:
+      - 'ibmiuser1'
+      - 'ibmiuser2'
+      - 'ibmiuser3'
+    fields:
+      - {'name':'status', 'expect':['*enabled']}
+      - {'name':'NO_PASSWORD_INDICATOR', 'expect':['no']}
+      - {'name':'SPECIAL_AUTHORITIES', 'expect': ['*JOBCTL', '*SAVSYS']}
 '''
 
 
@@ -217,7 +216,7 @@ from ansible.module_utils.basic import AnsibleModule
 from ansible_collections.ibm.power_ibmi.plugins.module_utils.ibmi import ibmi_util
 from ansible_collections.ibm.power_ibmi.plugins.module_utils.ibmi import ibmi_module as imodule
 
-__ibmi_module_version__ = "3.2.1"
+__ibmi_module_version__ = "3.3.0"
 
 parmname_array = [
     'SIGN_ON_ATTEMPTS_NOT_VALID',